requirements.txt
html2image
Pillow
selenium
pandas
openpyxl